Added mock Services tests, security, docker and migrations

+++ src/main/java/ru/chuchkalov/taskmanager/service/TaskService.java
@@ -3,12 +3,15 @@ package ru.chuchkalov.taskmanager.service;
 import lombok.AllArgsConstructor;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import ru.chuchkalov.taskmanager.dto.TaskRequestDTO;
 import ru.chuchkalov.taskmanager.dto.TaskResponseDTO;
 import ru.chuchkalov.taskmanager.entity.Task;
 import ru.chuchkalov.taskmanager.exception.EntityNotFoundException;
+import ru.chuchkalov.taskmanager.exception.AccessDeniedException;
 import ru.chuchkalov.taskmanager.mapper.TaskMapper;
 import ru.chuchkalov.taskmanager.repository.TaskRepository;
 import ru.chuchkalov.taskmanager.repository.UserRepository;
@@ -70,8 +73,19 @@ public class TaskService {
     }
 
     public void deleteTask(Long id) {
-        taskRepository.findById(id).orElseThrow(
+
+        Authentication auth = SecurityContextHolder.getContext().getAuthentication();
+
+        Task task = taskRepository.findById(id).orElseThrow(
                 () -> new EntityNotFoundException("Task with ID " + id + " not found"));
+
+        boolean isAdmin = auth.getAuthorities().stream()
+                .anyMatch(a -> a.getAuthority().equals("ROLE_USER"));
+        boolean isOwner = task.getUser().getUsername().equals(auth.getName());
+
+        if (!isAdmin && !isOwner) {
+            throw new AccessDeniedException("You cannot delete task with ID " + task.getId());
+        }
         taskRepository.deleteById(id);
     }
